# backend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ISARD Code Battle",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthRead"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/challenges/available": {
      "get": {
        "summary": "Available",
        "operationId": "available_api_challenges_available_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChallengeRead"
                  },
                  "type": "array",
                  "title": "Response Available Api Challenges Available Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/challenges/{id}/package": {
      "get": {
        "summary": "Package",
        "operationId": "package_api_challenges__id__package_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/zip": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/submissions": {
      "post": {
        "summary": "Submit",
        "operationId": "submit_api_submissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/organizer/dashboard": {
      "get": {
        "summary": "Dashboard",
        "operationId": "dashboard_api_organizer_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/organizer/advance": {
      "post": {
        "summary": "Advance",
        "operationId": "advance_api_organizer_advance_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/organizer/state": {
      "patch": {
        "summary": "Change State",
        "operationId": "change_state_api_organizer_state_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ChallengeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["id", "position", "title", "description"],
        "title": "ChallengeRead"
      },
      "CompetitionStateRead": {
        "properties": {
          "current_challenge": {
            "type": "integer",
            "title": "Current Challenge"
          },
          "individual_progress_enabled": {
            "type": "boolean",
            "title": "Individual Progress Enabled"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused"
          }
        },
        "type": "object",
        "required": ["current_challenge", "individual_progress_enabled", "paused"],
        "title": "CompetitionStateRead"
      },
      "Dashboard": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/CompetitionStateRead"
          },
          "global_challenge_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Challenge Title"
          },
          "can_advance": {
            "type": "boolean",
            "title": "Can Advance"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantStatus"
            },
            "type": "array",
            "title": "Participants"
          }
        },
        "type": "object",
        "required": ["state", "global_challenge_title", "can_advance", "participants"],
        "title": "Dashboard"
      },
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": ["detail"]
      },
      "HealthRead": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "HealthRead"
      },
      "ParticipantRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["id", "name", "created_at"],
        "title": "ParticipantRead"
      },
      "ParticipantStatus": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "challenge_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Position"
          },
          "challenge_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Title"
          },
          "status": {
            "type": "string",
            "enum": ["not_started", "working", "completed"],
            "title": "Status"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          }
        },
        "type": "object",
        "required": ["id", "name", "challenge_position", "challenge_title", "status", "score"],
        "title": "ParticipantStatus"
      },
      "StatePatch": {
        "properties": {
          "paused": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused"
          },
          "individual_progress_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Progress Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StatePatch"
      },
      "SubmissionCreate": {
        "properties": {
          "challenge_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Challenge Id"
          },
          "files": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "maxProperties": 100,
            "minProperties": 1,
            "title": "Files"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["challenge_id", "files"],
        "title": "SubmissionCreate"
      },
      "SubmissionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "participant_id": {
            "type": "integer",
            "title": "Participant Id"
          },
          "challenge_id": {
            "type": "integer",
            "title": "Challenge Id"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "accepted", "rejected"],
            "title": "Status"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["id", "participant_id", "challenge_id", "status", "feedback", "created_at"],
        "title": "SubmissionRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}